+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.0.6
 * Backported `stateSerializer` prop in `TanokInReact`
+* Fixed `@tanokComponent` passing wrong subName to `StreamWrapper.subWithMeta`
 
 ## 2.0.5
 * Fixed force updating of `TanokInReact` on state change